renames to OnboardingCTAContainer
