+++ src/components/OnboardingCTAContainer.svelte
